feat: show services in profile and direction link website project-detail and update ui post in profile
