+++ src/assets/icons/IconShareProject.tsx
@@ -1,0 +1,16 @@
+export default function IconShareProject() {
+  return (
+    <svg
+      width={16}
+      height={16}
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M11.9991 9.33333C11.6052 9.33593 11.2168 9.42576 10.8618 9.59636C10.5068 9.76696 10.1939 10.0141 9.94579 10.32L6.54579 8.75333C6.70565 8.26383 6.70565 7.73616 6.54579 7.24666L9.94579 5.68C10.3469 6.16397 10.906 6.49071 11.5246 6.60255C12.1431 6.7144 12.7813 6.60415 13.3264 6.29126C13.8716 5.97838 14.2887 5.48298 14.5042 4.89249C14.7196 4.302 14.7196 3.65438 14.504 3.06393C14.2884 2.47348 13.8712 1.97816 13.326 1.66537C12.7808 1.35259 12.1426 1.24246 11.5241 1.35441C10.9056 1.46636 10.3465 1.79321 9.94548 2.27726C9.54449 2.76131 9.32738 3.37145 9.33246 3.99999C9.33446 4.15882 9.35007 4.31717 9.37912 4.47333L5.85912 6.09333C5.48384 5.72638 5.00862 5.47834 4.49299 5.38028C3.97737 5.28222 3.44427 5.33851 2.96049 5.54208C2.47671 5.74566 2.06377 6.08747 1.77339 6.5247C1.48302 6.96193 1.32812 7.47513 1.32812 8C1.32812 8.52486 1.48302 9.03806 1.77339 9.47529C2.06377 9.91252 2.47671 10.2543 2.96049 10.4579C3.44427 10.6615 3.97737 10.7178 4.49299 10.6197C5.00862 10.5217 5.48384 10.2736 5.85912 9.90666L9.37912 11.5267C9.35007 11.6828 9.33446 11.8412 9.33246 12C9.33246 12.5274 9.48885 13.043 9.78187 13.4815C10.0749 13.92 10.4914 14.2618 10.9786 14.4637C11.4659 14.6655 12.0021 14.7183 12.5194 14.6154C13.0366 14.5125 13.5118 14.2586 13.8847 13.8856C14.2577 13.5127 14.5117 13.0375 14.6146 12.5202C14.7174 12.003 14.6646 11.4668 14.4628 10.9795C14.261 10.4922 13.9192 10.0758 13.4806 9.78274C13.0421 9.48973 12.5265 9.33333 11.9991 9.33333ZM11.9991 2.66666C12.2628 2.66666 12.5206 2.74486 12.7399 2.89137C12.9591 3.03788 13.13 3.24612 13.231 3.48975C13.3319 3.73339 13.3583 4.00147 13.3068 4.26012C13.2554 4.51876 13.1284 4.75633 12.9419 4.9428C12.7555 5.12927 12.5179 5.25626 12.2592 5.30771C12.0006 5.35916 11.7325 5.33275 11.4889 5.23183C11.2452 5.13092 11.037 4.96002 10.8905 4.74076C10.744 4.52149 10.6658 4.2637 10.6658 3.99999C10.6658 3.64637 10.8063 3.30723 11.0563 3.05719C11.3064 2.80714 11.6455 2.66666 11.9991 2.66666ZM3.99912 9.33333C3.73541 9.33333 3.47763 9.25513 3.25836 9.10862C3.0391 8.96211 2.8682 8.75387 2.76728 8.51024C2.66637 8.2666 2.63996 7.99852 2.69141 7.73987C2.74286 7.48123 2.86984 7.24366 3.05631 7.05719C3.24278 6.87072 3.48036 6.74373 3.739 6.69228C3.99764 6.64083 4.26573 6.66724 4.50937 6.76816C4.753 6.86907 4.96124 7.03997 5.10775 7.25923C5.25426 7.4785 5.33246 7.73629 5.33246 8C5.33246 8.35362 5.19198 8.69276 4.94193 8.9428C4.69188 9.19285 4.35275 9.33333 3.99912 9.33333ZM11.9991 13.3333C11.7354 13.3333 11.4776 13.2551 11.2584 13.1086C11.0391 12.9621 10.8682 12.7539 10.7673 12.5102C10.6664 12.2666 10.64 11.9985 10.6914 11.7399C10.7429 11.4812 10.8698 11.2437 11.0563 11.0572C11.2428 10.8707 11.4804 10.7437 11.739 10.6923C11.9976 10.6408 12.2657 10.6672 12.5094 10.7682C12.753 10.8691 12.9612 11.04 13.1077 11.2592C13.2543 11.4785 13.3325 11.7363 13.3325 12C13.3325 12.3536 13.192 12.6928 12.9419 12.9428C12.6919 13.1929 12.3527 13.3333 11.9991 13.3333Z"
+        fill="#82EBFF"
+      />
+    </svg>
+  );
+}
