+++ src/modules/profile/view/profile.tsx
@@ -1,34 +1,18 @@
 "use client";
-import {
-  IconBlue,
-  IconEdit,
-  IconPointProfile,
-  IconStarNormal,
-  IconVerify,
-  IconStar
-} from "@/assets/icons";
-import { ButtonPrimary } from "@/components/ButtonCustom";
-import { SOCIAL } from "@/constant/social";
-import { useBoolean } from "@/hooks/useBoolean";
-import { Box, Divider, TextField, Typography } from "@mui/material";
-import Image from "next/image";
-import { useEffect, useState } from "react";
+
+import { Divider } from "@mui/material";
 import styled from "styled-components";
 import Experience from "../components/Experiences";
 import PostUser from "../components/PostUser";
 import Services from "../components/services";
-import { getMyProfile } from "./../../../services/index";
-import EditProfile from "@/components/EditProfile";
-import LoadingSkeleton from "@/components/LoadingSkeleton";
 import PersonSkeleton from "@/components/Skeleton/PersonSkeleton";
 import PostSkeleton from "@/components/Skeleton/PostSkeleton";
 import OverviewSkeleton from "@/components/Skeleton/OverviewSkeleton";
 import ServicesSkeleton from "@/components/Skeleton/ServicesSkeleton";
 import { useServicesContext } from "@/modules/services/context/ServicesContext";
-import { useProfileContext } from "@/contexts/ProfileContext";
 import { useMyProfileContext } from "@/contexts/MyProfileConext";
-import Chips from "@/components/Chip";
 import CompletedProject from "../components/CompletedProject";
+import PersonalUserProfile from "../components/PersonalUserProfile";
 
 export interface IUserProfileProps {
   widthNotData?: boolean;
@@ -36,185 +20,6 @@ export interface IUserProfileProps {
 
 const IMG_NFT =
   "https://images.pexels.com/photos/842711/pexels-photo-842711.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1";
-
-const Overview = ({ overview }: any) => {
-  const openModal = useBoolean();
-  return overview ? (
-    <StyleOverview>
-      <StyleLeft>
-        <StyleTitle>Overview</StyleTitle>
-        <PrimaryTitleLeft>
-          <StyleContentOverview>
-            <StyleDesOverview>Primary Job Title</StyleDesOverview>
-            <StyleSubTitle>Fashion KOL</StyleSubTitle>
-          </StyleContentOverview>
-          <StyleContentOverview>
-            <StyleDesOverview>Primary Job Title</StyleDesOverview>
-            <StyleSubTitle>Fashion KOL</StyleSubTitle>
-          </StyleContentOverview>
-        </PrimaryTitleLeft>
-        <PrimaryTitleLeft>
-          <StyleContentOverview>
-            <StyleDesOverview>Primary Job Title</StyleDesOverview>
-            <StyleSubTitle>Fashion KOL</StyleSubTitle>
-          </StyleContentOverview>
-          <StyleContentOverview>
-            <StyleDesOverview>Primary Job Title</StyleDesOverview>
-            <StyleSubTitle>Fashion KOL</StyleSubTitle>
-          </StyleContentOverview>
-          <StyleContentOverview>
-            <StyleDesOverview>Primary Job Title</StyleDesOverview>
-            <StyleSubTitle>
-              Fashion KOL Fashion KOL Fashion KOL Fashion KOL
-            </StyleSubTitle>
-          </StyleContentOverview>
-        </PrimaryTitleLeft>
-      </StyleLeft>
-      <StyleRight>
-        <PrimaryTitleRight>
-          <StyleContentOverview>
-            <StyleDesOverview>Twitter</StyleDesOverview>
-            <StyleSubTitle>
-              {overview?.twitterInfo?.followers ?? 0}
-            </StyleSubTitle>
-          </StyleContentOverview>
-          <StyleContentOverview>
-            <StyleDesOverview>Primary per post</StyleDesOverview>
-            <StyleSubTitle>$12,450</StyleSubTitle>
-          </StyleContentOverview>
-        </PrimaryTitleRight>
-        {/* <ButtonPrimary onClick={() => openModal?.onTrue()}>
-          <Typography sx={{ p: "8px 0" }}>Hire Me</Typography>
-        </ButtonPrimary> */}
-      </StyleRight>
-      {openModal?.value && (
-        <div style={{ width: "300ox", height: "300px" }}>
-          {/* <ModalRequestCollaboration openHireMe={openModal.onFalse} /> */}
-        </div>
-      )}
-    </StyleOverview>
-  ) : (
-    <ContentNotData>
-      <StyleTitle>OverView</StyleTitle>
-      <DescriptionNotData>
-        {`You don't have any work Overview yet.`}
-      </DescriptionNotData>
-    </ContentNotData>
-  );
-};
-
-const Personal = ({ dataPersonal, resetPage }: any) => {
-  const isOpenEditProfile = useBoolean();
-  const openModal = useBoolean();
-  return (
-    <StylePersonal>
-      <StylePersonalLeft>
-        <StyleImage
-          src={dataPersonal?.twitterInfo?.avatar ?? IMG_NFT}
-          alt="avatar profile"
-          width={220}
-          height={220}
-        />
-        <StyleContentUser>
-          <StyleTitle>
-            {dataPersonal?.fullName ?? "User Name"}
-            {dataPersonal?.twitterInfo?.verificationStatus && <IconVerify />}
-          </StyleTitle>
-          <PointProfile>
-            <IconPointProfile />
-            {dataPersonal?.twitterInfo?.totalPoints ?? 0}
-          </PointProfile>
-          <StyleContentFlex>
-            <StyleContentUser>
-              <StyleUserDes>{dataPersonal?.bio ?? "Data null"}</StyleUserDes>
-              <StyleUserDes>Influencer</StyleUserDes>
-              <StyleUserSocial>Social</StyleUserSocial>
-              <StyleIcons>
-                {dataPersonal?.socialProfiles?.map(
-                  (item: any, index: number) => SOCIAL[item?.social] ?? <></>,
-                )}
-              </StyleIcons>
-            </StyleContentUser>
-            <StyleContentUser>
-              <StyleTotal>
-                <StyleDesOverview>Total Achievements:</StyleDesOverview>
-                <StyleSubTitle>32</StyleSubTitle>
-              </StyleTotal>
-              <StyleTotal>
-                <StyleDesOverview>Review:</StyleDesOverview>
-                <StyleSubTitle>{dataPersonal?.review}</StyleSubTitle>
-              </StyleTotal>
-              <StyleIcons>
-                <IconStar />
-                <IconStar />
-                <IconStar />
-                <IconStar />
-                <IconStar />
-              </StyleIcons>
-            </StyleContentUser>
-            <StyleContentUser>
-              <StyleDesOverview>Location</StyleDesOverview>
-              <StyleSubTitle>{dataPersonal.location ? dataPersonal.location : "-"}</StyleSubTitle>
-            </StyleContentUser>
-          </StyleContentFlex>
-          {/* {dataPersonal.bio && <StyleUserDes>{dataPersonal.bio}</StyleUserDes>}
-          <div style={{ display: "flex", gap: "16px" }}>
-            <StyleUserDes>Influencer</StyleUserDes>
-            <StyleContentOverview>
-              <StyleDesOverview>Location</StyleDesOverview>
-              <StyleSubTitle>{dataPersonal.location ? dataPersonal.location : "-"}</StyleSubTitle>
-            </StyleContentOverview>
-          </div>
-          <StyleUserSocial>Social</StyleUserSocial>
-          <StyleIcons>
-            {dataPersonal?.socialProfiles?.map(
-              (item: any, index: number) => SOCIAL[item?.social] ?? <></>,
-            )}
-          </StyleIcons> */}
-        </StyleContentUser>
-      </StylePersonalLeft>
-      <StylePersonalRight>
-        <StyleButtons>
-          <StyleButtonTitle onClick={isOpenEditProfile.onTrue}>
-            <IconEdit />
-            <div>Edit</div>
-          </StyleButtonTitle>
-          <StyleButtonTitle>
-            <IconBlue />
-            Share
-          </StyleButtonTitle>
-          <StyleButtonTitle>
-            <IconStarNormal />
-            Follow <strong>{dataPersonal?.fullName}</strong>
-          </StyleButtonTitle>
-        </StyleButtons>
-        <StyleChips>
-          {dataPersonal?.tags?.map((listTag: string, index: number) => (
-            <Chips key={index} label={listTag} color="secondary" sx={{ color: "#25002D", backgroundColor: "#F6CCFF" }} />
-          ))}
-        </StyleChips>
-        <StyleContentUser style={{ paddingTop: "12px" }}>
-          <StyleTotal>
-            <StyleDesOverview>X Followers:</StyleDesOverview>
-            <StyleSubTitle>{dataPersonal?.twitterInfo?.followers}</StyleSubTitle>
-          </StyleTotal>
-        </StyleContentUser>
-        <ButtonPrimary onClick={() => openModal?.onTrue()}>
-          <Typography sx={{ p: "8px 0" }}>
-            DM to {dataPersonal?.fullName}
-          </Typography>
-        </ButtonPrimary>
-      </StylePersonalRight>
-      {isOpenEditProfile.value && (
-        <EditProfile
-          resetModal={isOpenEditProfile.onFalse}
-          dataPersonal={dataPersonal}
-          resetPage={resetPage}
-        />
-      )}
-    </StylePersonal>
-  );
-};
 
 export default function UserProfile(props: IUserProfileProps) {
   const { dataPopularServices } = useServicesContext();
@@ -226,7 +31,10 @@ export default function UserProfile(props: IUserProfileProps) {
       {isLoading ? (
         <PersonSkeleton />
       ) : dataPersonal ? (
-        <Personal dataPersonal={dataPersonal} resetPage={() => fetchData()} />
+        <PersonalUserProfile
+          dataPersonal={dataPersonal}
+          resetPage={() => fetchData()}
+        />
       ) : (
         <PersonSkeleton />
       )}
@@ -278,20 +86,20 @@ export default function UserProfile(props: IUserProfileProps) {
                 dataPopularServices={dataPopularServices}
                 services={dataPersonal}
               />
-              <Post>
-                <StyleTitle>Recent posts</StyleTitle>
-                <Posts widthNotData={dataPosts?.length > 0}>
-                  {dataPosts?.length > 0 ? (
-                    dataPosts.map((item: any, index: number) => (
-                      <>
-                        <PostUser item={item} />
-                      </>
-                    ))
-                  ) : (
-                    <PostNotData>{`You haven't made any posts yet.`}</PostNotData>
-                  )}
-                </Posts>
-              </Post>
+              <StyleBox>
+                <Post>
+                  <StyleTitle>Recent posts</StyleTitle>
+                  <Posts widthNotData={dataPosts?.length > 0}>
+                    {dataPosts?.length > 0 ? (
+                      dataPosts.map((item: any, index: number) => (
+                        <PostUser key={index} item={item} />
+                      ))
+                    ) : (
+                      <PostNotData>{`You haven't made any posts yet.`}</PostNotData>
+                    )}
+                  </Posts>
+                </Post>
+              </StyleBox>
               {/* <Divider sx={{ borderColor: "#B9B9B9 " }} /> */}
             </>
           )}
@@ -301,36 +109,12 @@ export default function UserProfile(props: IUserProfileProps) {
   );
 }
 
-
-const StyleChips = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 12px;
-  padding: 4px 0;
-  color: #ffff !important;
-`;
-
-
-const StyleTotal = styled.div`
-  display: flex;
-  align-items: center;
-  width: 215px;
-  gap: 10px;
-`;
-
-const StyleContentFlex = styled.div`
-  display: flex;
-  align-items: start;
-  flex-wrap: wrap;
-  gap: 36px;
-`;
-
 const Post = styled.div`
   padding: 24px 14px;
   width: 100%;
   overflow-x: hidden;
   margin-top: 24px;
-`
+`;
 
 const Content = styled.div`
   display: flex;
@@ -342,31 +126,6 @@ const Content = styled.div`
 
 const ContentRight = styled.div`
   width: 100%;
-  @media (max-width: 768px) {
-    width: 100%;
-  }
-`;
-
-const ContentNotData = styled.div`
-  padding: 20px 15px;
-`;
-
-const DescriptionNotData = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 300px;
-  color: #f23581;
-`;
-
-const PostLeft = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  width: 30%;
-  padding: 12px;
   @media (max-width: 768px) {
     width: 100%;
   }
@@ -389,223 +148,15 @@ const PostNotData = styled.div`
   color: #f23581;
 `;
 
-const StyleButtons = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 24px;
-  @media (max-width: 520px) {
-    flex-wrap: no-wrap;
-    gap: 4px;
-  }
-`;
-const StyleButtonTitle = styled.div`
-  padding: 4px 8px;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  background-color: #393939;
-  cursor: pointer;
-  color: #b9b9b9;
-  border-radius: 6px;
-`;
 const StyleContainer = styled.div`
   background-color: #292d32;
   color: #ffffff;
 `;
-const PrimaryTitleLeft = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 60px;
-  padding: 24px 0;
-`;
 
-const PrimaryTitleRight = styled.div`
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  gap: 60px;
-  padding: 24px 0;
-`;
-
-const StyleDesOverview = styled.div`
-  font-size: 16px;
-  font-weight: 500;
-  line-height: 21px;
-  color: #b9b9b9;
-`;
-const StyleContentOverview = styled.div``;
-const StyleSubTitle = styled.div`
-  padding-top: 8px;
-  font-size: 18px;
-  font-weight: 700;
-  line-height: 24px;
-  color: #82ebff;
-`;
-const StylePersonal = styled.div`
-  display: flex;
-  align-items: flex-start;
-  padding: 24px 14px;
-  @media (max-width: 520px) {
-    flex-wrap: wrap;
-    padding: 8px;
-    justify-content: center;
-  }
-`;
-const StyleImage = styled(Image)`
-  border: 2px solid #fff;
-  border-radius: 50%;
-`;
-const StylePersonalLeft = styled.div`
-  width: 60%;
-  display: flex;
-  align-items: center;
-  gap: 32px;
-  @media (max-width: 520px) {
-    width: 100%;
-    justify-content: center;
-  }
-`;
-const StyleContentUser = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-`;
 const StyleTitle = styled.div`
   font-size: 40px;
   line-height: 51px;
   font-weight: 700;
-`;
-const PointProfile = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-`;
-const StyleUserDes = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 4px 12px;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 20px;
-  background-color: #393939;
-  border-radius: 12px;
-  color: #ffd7f4;
-  width: fit-content;
-`;
-const StyleUserSocial = styled.div`
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 18px;
-  color: #b9b9b9;
-`;
-const StyleIcons = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 8px;
-`;
-
-const StylePersonalRight = styled.div`
-  margin-left: 50px;
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  width: 40%;
-  @media (max-width: 520px) {
-    width: 100%;
-    margin-left: 0px;
-    margin-top: 30px;
-  }
-`;
-const StyleOverview = styled.div`
-  display: flex;
-  justify-content: space-around;
-  padding: 16px 20px;
-  @media (max-width: 520px) {
-    flex-direction: column;
-    align-items: center;
-    padding: 24px 14px;
-  }
-`;
-const StyleLeft = styled.div``;
-
-const StyleRight = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 80%;
-`;
-
-const StyleModalBox = styled(Box)`
-  position: fixed;
-  z-index: 99999999;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 400;
-  background-color: #2e2e2e;
-  border: 1px solid #2e2e2e;
-  color: #fff;
-  box-shadow: 24;
-  padding: 0;
-`;
-
-const StyleTop = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  margin-top: 10px;
-`;
-
-const StyleButtonClose = styled.div`
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  cursor: pointer;
-  :hover {
-    opacity: 0.8;
-  }
-`;
-
-const StyleBottom = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-  /* gap: 5px; */
-  flex-wrap: nowrap;
-  background-color: #252525;
-  border-radius: 10px;
-  width: 100%;
-  padding: 20px 50px;
-`;
-
-const StyleLabel = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  margin-bottom: 8px;
-`;
-
-const StyleBottomSubmit = styled.div`
-  width: 100%;
-  margin-top: 20px;
-`;
-
-const StyleBottomGotIt = styled.div`
-  width: 100%;
-  margin: 20px 0;
-`;
-
-const StyleInput = styled(TextField)`
-  border-radius: 20px;
-  color: #fff;
-  background-color: #353535;
-  border: 0px #353535 solid;
 `;
 
 export const StyleError = styled.p`
@@ -615,11 +166,8 @@ export const StyleError = styled.p`
   white-space: normal;
 `;
 
-const StyleRequest = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-  align-items: center;
-  color: #f23581;
-  margin-top: 15px;
+const StyleBox = styled.div`
+  padding: 24px 14px;
+  width: 100%;
+  overflow-x: hidden;
 `;
